+++ packages/editor-bricks-helper/src/index.ts
@@ -14,4 +14,7 @@ export * from "./hooks/useBuilderNodeMountPoints";
 export * from "./hooks/useBuilderParentNode";
 export * from "./hooks/useCanDrop";
 export * from "./hooks/useRouteList";
+export * from "./hooks/useHoverNodeUid";
+export * from "./hooks/useShowRelatedNodesBasedOnEvents";
 export * from "./processors/getSortedIdsAfterDropped";
+export * from "./assertions";