+++ packages/brick-types/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.20.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-types@2.20.0...@next-core/brick-types@2.20.1) (2021-03-24)
+
+**Note:** Version bump only for package @next-core/brick-types
+
+
+
+
+
+# [2.20.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-types@2.19.1...@next-core/brick-types@2.20.0) (2021-03-17)
+
+
+### Features
+
+* support context onChange ([09bd098](https://github.com/easyops-cn/next-core/commit/09bd098dba3b3319ecad39be0f7cfd877cb70c5a))
+
+
+
+
+
+## [2.19.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-types@2.19.0...@next-core/brick-types@2.19.1) (2021-03-08)
+
+**Note:** Version bump only for package @next-core/brick-types
+
+
+
+
+
 # [2.19.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-types@2.18.0...@next-core/brick-types@2.19.0) (2021-02-25)