+++ packages/brick-utils/CHANGELOG.md
@@ -3,6 +3,41 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.14.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-utils@2.14.0...@next-core/brick-utils@2.14.1) (2021-03-24)
+
+**Note:** Version bump only for package @next-core/brick-utils
+
+
+
+
+
+# [2.14.0](https://github.com/easyops-cn/next-core/compare/@next-core/brick-utils@2.13.3...@next-core/brick-utils@2.14.0) (2021-03-18)
+
+
+### Features
+
+* **evaluations:** support URLSearchParams ([3433233](https://github.com/easyops-cn/next-core/commit/3433233446c52778e904d7a96ccb8d6d33419043))
+
+
+
+
+
+## [2.13.3](https://github.com/easyops-cn/next-core/compare/@next-core/brick-utils@2.13.2...@next-core/brick-utils@2.13.3) (2021-03-17)
+
+**Note:** Version bump only for package @next-core/brick-utils
+
+
+
+
+
+## [2.13.2](https://github.com/easyops-cn/next-core/compare/@next-core/brick-utils@2.13.1...@next-core/brick-utils@2.13.2) (2021-03-08)
+
+**Note:** Version bump only for package @next-core/brick-utils
+
+
+
+
+
 ## [2.13.1](https://github.com/easyops-cn/next-core/compare/@next-core/brick-utils@2.13.0...@next-core/brick-utils@2.13.1) (2021-03-04)