+++ packages/create-next-repo/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.3.4](https://github.com/easyops-cn/next-core/compare/@next-core/create-next-repo@1.3.3...@next-core/create-next-repo@1.3.4) (2021-03-08)
+
+**Note:** Version bump only for package @next-core/create-next-repo
+
+
+
+
+
 ## [1.3.3](https://github.com/easyops-cn/next-core/compare/@next-core/create-next-repo@1.3.2...@next-core/create-next-repo@1.3.3) (2021-02-22)
 
 **Note:** Version bump only for package @next-core/create-next-repo